==> lab_04/0-weather_data_collection.ipynb <==
# Collecting weather data from an API

This notebook contains the code that was used to collect the data for this lab. Note that if you overwrite the data that came with this lab by saving the data you collect here, your results in the remaining notebooks may not match the course due to changes in the NCEI API's data.

## About the data
In this notebook, we will be collecting daily weather data from the [National Centers for Environmental Information (NCEI) API](https://www.ncdc.noaa.gov/cdo-web/webservices/v2). We will use the Global Historical Climatology Network - Daily (GHCND) dataset; see the documentation [here](https://www1.ncdc.noaa.gov/pub/data/cdo/documentation/GHCND_documentation.pdf).

*Note: The NCEI is part of the National Oceanic and Atmospheric Administration (NOAA) and, as you can see from the URL for the API, this resource was created when the NCEI was called the NCDC. Should the URL for this resource change in the future, you can search for "NCEI weather API" to find the updated one.*

## Using the NCEI API
Request your token [here](https://www.ncdc.noaa.gov/cdo-web/token) and then paste it below.

import requests

def make_request(endpoint, payload=None):
    """
    Make a request to a specific endpoint on the weather API
    passing headers and optional payload.
    
    Parameters:
        - endpoint: The endpoint of the API you want to 
                    make a GET request to.
        - payload: A dictionary of data to pass along 
                   with the request.
    
    Returns:
        Response object.
    """
    return requests.get(
        f'https://www.ncdc.noaa.gov/cdo-web/api/v2/{endpoint}',
        headers={
            'token': 'PASTE_YOUR_TOKEN_HERE'
        },
        params=payload
    )

## Collect All Data Points for 2018 In NYC (Various Stations)
We can make a loop to query for all the data points one day at a time, providing updates using `IPython.display`. Here we create a list of all the results:

import datetime

from IPython import display # for updating the cell dynamically

current = datetime.date(2018, 1, 1)
end = datetime.date(2019, 1, 1)

results = []

while current < end:
    # update the cell with status information
    display.clear_output(wait=True)
    display.display(f'Gathering data for {str(current)}')
    
    response = make_request(
        'data', 
        {
            'datasetid': 'GHCND', # Global Historical Climatology Network - Daily (GHCND) dataset
            'locationid': 'CITY:US360019', # NYC
            'startdate': current,
            'enddate': current,
            'units': 'metric',
            'limit': 1000 # max allowed
        }
    )

    if response.ok:
        # we extend the list instead of appending to avoid getting a nested list
        results.extend(response.json()['results'])

    # update the current date to avoid an infinite loop
    current += datetime.timedelta(days=1)

Now, we can create a dataframe with all this data. Notice there are multiple stations with values for each `datatype` on a given day. We don't know what the stations are, but we can look them up and add them to the data:

import pandas as pd

df = pd.DataFrame(results)
df.head()

Save this data to a file:

df.to_csv('data/nyc_weather_2018.csv', index=False)

and write it to the database:

import sqlite3

with sqlite3.connect('data/weather.db') as connection:
    df.to_sql(
        'weather', connection, index=False, if_exists='replace'
    )

For learning about merging dataframes, we will also get the data mapping station IDs to information about the station:

response = make_request(
    'stations', 
    {
        'datasetid': 'GHCND', # Global Historical Climatology Network - Daily (GHCND) dataset
        'locationid': 'CITY:US360019', # NYC
        'limit': 1000 # max allowed
    }
)

stations = pd.DataFrame(response.json()['results'])[['id', 'name', 'latitude', 'longitude', 'elevation']]
stations.to_csv('data/weather_stations.csv', index=False)

with sqlite3.connect('data/weather.db') as connection:
    stations.to_sql(
        'stations', connection, index=False, if_exists='replace'
    )

<hr>
<div>
    <a href="../lab_03/5-handling_data_issues.ipynb">
        <button>&#8592; Lab 8</button>
    </a>
    <a href="./1-querying_and_merging.ipynb">
        <button style="float: right;">Next Notebook &#8594;</button>
    </a>
</div>
<hr>
